+++ core/src/ClassicalBody.cpp
@@ -3,28 +3,27 @@
 */
 #include <ClassicalBody.hh>
 #include <iostream>
+#include <Units.hh>
 
 
-void ClassicalBody::update(double dt) {
+void ClassicalBody::update(double dt, Units& units) {
 
     // update position using veloctity
     glm::dvec3 velocity = this->getVelocity();
     glm::dvec3 position = this->getPosition();
     glm::dvec3 acceleration = this->getAcceleration();
-    double G = 6.67430e-11; // gravitational constant
-    double M = 5.972e8;
+    double G = units.G; // gravitational constant
+    double M = 1.989e30/units.massScale;
 
     glm::dvec3 newAcceleration = -G * M * position / glm::length(position) / glm::length(position) / glm::length(position);
+    // std::cout << glm::length(position) << " " << G * M << glm::length(newAcceleration) << " a"<< std::endl;
 
     // use leapfrog integration to update velocity and position
     glm::dvec3 halfVelocity = velocity + 0.5 * acceleration * dt;
     glm::dvec3 newPosition = position + halfVelocity * dt;
     glm::dvec3 newVelocity = halfVelocity + 0.5 * newAcceleration * dt;
 
-    // std::cout << "Acceleration: " << newAcceleration.x << ", " << newAcceleration.y << ", " << newAcceleration.z << std::endl;
-    // std::cout << "Velocity: " << newVelocity.x << ", " << newVelocity.y << ", " << newVelocity.z << std::endl;
-    // std::cout << "Position: " << newPosition.x << ", " << newPosition.y << ", " << newPosition.z << std::endl;
-
+    // std::cout << glm::length(newVelocity) << std::endl;
     this->setPosition(newPosition);
     this->setVelocity(newVelocity);
     this->setAcceleration(newAcceleration);
@@ -39,10 +38,10 @@ void ClassicalBody::addTrailPoint(const glm::dvec3& position) {
 }
 
 
-std::vector<glm::vec4> ClassicalBody::getRenderData(const glm::dvec3& cameraPos) {
+std::vector<glm::vec4> ClassicalBody::getRenderData(const glm::dvec3& cameraPos, Units& units) {
 
     // get position and speed for rendering
-    glm::vec3 relativePosition = static_cast<glm::vec3>(getPosition() - cameraPos);
+    glm::vec3 relativePosition = static_cast<glm::vec3>((getPosition() - cameraPos));
     float speed = getSpeed();
 
     return {glm::vec4(relativePosition, speed)};